Set properties to private where possible
