--- Cineaste/ViewController/MovieNight/MovieNightViewController+UITableView.swift
@@ -1,21 +1,0 @@
-//
-//  MovieNightViewController+UITableView.swift
-//  Cineaste
-//
-//  Created by Felizia Bernutz on 21.07.18.
-//  Copyright © 2018 spacepandas.de. All rights reserved.
-//
-
-import UIKit
-
-extension MovieNightViewController: UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return nearbyMessages.count
-    }
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell: MovieNightUserCell = tableView.dequeueCell(identifier: MovieNightUserCell.identifier)
-        cell.configure(with: nearbyMessages[indexPath.row])
-        return cell
-    }
-}
